requirements.txt
langchain==0.0.335
faiss-cpu==1.7.4
huggingface-hub==0.19.0
sentence-transformers==2.2.2
pypdf==3.17.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
httpx==0.27.2
azure-storage-blob==12.18.3
flask==2.3.3
beautifulsoup4==4.12.2
trafilatura==1.6.1
streamlit==1.29.0
uuid==1.30